wallet_api/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12.0
psycopg2-binary>=2.9.0
gunicorn>=20.1
django-environ
